backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
openai>=1.12.0
python-multipart==0.0.9
pdfplumber==0.11.0
python-dotenv==1.0.1
pydantic==2.6.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
google-auth==2.28.0
requests==2.31.0
trafilatura>=1.6.0
